backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
cryptography
alembic
pydantic
python-dotenv
azure-storage-blob
python-multipart